services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg[binary]==3.2.3
pydantic==2.10.6
python-dotenv==1.0.1
PyJWT==2.10.1
argon2-cffi==23.1.0
